day9/Online_Vegetable/requirements.txt
python-multipart
bcrypt==4.0.1
email-validator
fastapi
uvicorn
motor
pydantic
passlib[bcrypt]
python-jose
pytest
httpx
pydantic-settings